dom/media/webaudio/test/test_audioDestinationNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test AudioDestinationNode as EventTarget</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var ac = new AudioContext()
ac.destination.addEventListener("foo", function() {
  ok(true, "Event received!");
  SimpleTest.finish();
});
ac.destination.dispatchEvent(new CustomEvent("foo"));

</script>
</pre>
</body>
</html>
